Added confirmation dialogs

+++ vmwiz-backend/router/vm.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"git.sos.ethz.ch/vsos/app.vsos.ethz.ch/vmwiz-backend/auth"
+	"git.sos.ethz.ch/vsos/app.vsos.ethz.ch/vmwiz-backend/confirmation"
 	"git.sos.ethz.ch/vsos/app.vsos.ethz.ch/vmwiz-backend/netcenter"
 	"git.sos.ethz.ch/vsos/app.vsos.ethz.ch/vmwiz-backend/proxmox"
 	"github.com/gorilla/mux"
@@ -17,7 +18,26 @@ import (
 
 func addAllVMRoutes(r *mux.Router) {
 
-	r.Methods("POST").Path("/api/vm/deleteByName").Subrouter().NewRoute().Handler(auth.CheckAuthenticated(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	r.Methods("POST").Path("/api/vm/deleteByName").Subrouter().NewRoute().Handler(auth.CheckAuthenticated(confirmation.ConfirmMiddleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if token := r.Context().Value(confirmation.ConfirmationTokenContextField); token != nil {
+			type response struct {
+				ConfirmationToken string `json:"confirmationToken"`
+			}
+
+			resp := response{
+				ConfirmationToken: *(token.(*string)),
+			}
+			respJSON, err := json.Marshal(resp)
+			if err != nil {
+				log.Printf("Error marshalling response: %v", err)
+				http.Error(w, "Failed to marshal response", http.StatusInternalServerError)
+				return
+			}
+			w.Header().Set("Content-Type", "application/json")
+			w.Write(respJSON)
+			return
+		}
+
 		type bodyS struct {
 			Name      string `json:"vmName"`
 			DeleteDNS bool   `json:"deleteDNS"`
@@ -83,5 +103,5 @@ func addAllVMRoutes(r *mux.Router) {
 		}
 
 		w.WriteHeader(http.StatusOK)
-	})))
+	}))))
 }
